Backend-Django/requirements.txt
Django==4.2.19
djangorestframework==3.14.0
drf_yasg==1.21.7
groq==0.18.0
joblib==1.2.0
numpy==1.22.4
pandas==1.4.4
Requests==2.32.3
scikit_learn==1.3.2
